implement auto login feature

+++ trunk/dailyreport/dailyreport/view/KPIViewController.m
@@ -116,7 +116,7 @@
     [self.contentView addSubview:[self createImageViewFromNamedImage:@"关键指标-分割线.png" frame:CGRectMake(0,157,320,3)]];
     [self.contentView addSubview:[self createImageViewFromNamedImage:@"票数.png" frame:CGRectMake(20,174,24,24)]];
     [self.contentView addSubview:[self createLabel:@"票数" frame:CGRectMake(50,160,100,49) textColor:@"#4a4a4a" font:18 backgroundColor:nil]];
-    self.docNumberLabel =[self createLabel:[NSString stringWithFormat:@"%0.2f",channel.docNumber.doubleValue] frame:CGRectMake(200,160,100,49) textColor:@"#7f7f7f" font:18 backgroundColor:nil];
+    self.docNumberLabel =[self createLabel:[NSString stringWithFormat:@"%d",channel.docNumber.intValue] frame:CGRectMake(200,160,100,49) textColor:@"#7f7f7f" font:18 backgroundColor:nil];
     [self.contentView addSubview:self.docNumberLabel];
     [self.contentView addSubview:[self createImageViewFromNamedImage:@"关键指标-分割线.png" frame:CGRectMake(0,217,320,3)]];
     [self.contentView addSubview:[self createImageViewFromNamedImage:@"附加.png" frame:CGRectMake(20,234,24,24)]];
@@ -142,7 +142,7 @@
 -(void) updateChannel:(Channel*) channel{
     self.dayAmountLabel.text =[NSString stringWithFormat:@"%d 万元",channel.dayAmount.intValue/10000];
     self.docNumberLabel.text = [NSString stringWithFormat:@"%d",channel.docNumber.intValue];
-    self.avgDocCountLabel.text = [NSString stringWithFormat:@"%d",channel.avgDocCount.intValue];
+    self.avgDocCountLabel.text = [NSString stringWithFormat:@"%0.2f",channel.avgDocCount.doubleValue];
     self.avgPriceLabel.text = [NSString stringWithFormat:@"%d 元",channel.avgPrice.intValue];
     self.apsLabel.text = [NSString stringWithFormat:@"%d 元",channel.aps.intValue];
 }
